feat(runner-tui): move register form from Runners tab to General tab
Registration binds the whole daemon to a cloud URL — that's a daemon-
level concern, not a per-runner one — so the form belongs on General.
The Runners tab now shows a placeholder pointing the user there when
no config.toml exists yet.
